PersonalProjects/NBA_Stats_LLM/ai-wrapper/requirements.txt
openai==1.3.7
google-generativeai==0.3.2
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
aiohttp==3.9.1
asyncio==3.4.3
json5==0.9.14
beautifulsoup4==4.12.2
lxml==4.9.3
